pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "django-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Aurélien Amorin",email = "amorin77184@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2,<5.3)",
    "djangorestframework (>=3.16,<3.17)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
]

[dependency-groups]
dev = [
    "black",
    "flake8 (>=7.3.0,<8.0.0)"
]

[tool.black]
line-length = 100

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
